#485: renaming query args and transformer
