+++ src/backend/src/services/teams.services.ts
@@ -1,7 +1,7 @@
 import { Team } from 'shared';
-import teamQueryArgs from '../prisma-query-args/team.query-args';
+import teamQueryArgs from '../prisma-query-args/teams.query-args';
 import prisma from '../prisma/prisma';
-import teamsTransformer from '../transformers/teams.transformer';
+import teamTransformer from '../transformers/teams.transformer';
 import { NotFoundException } from '../utils/errors.utils';
 
 export default class TeamsService {
@@ -11,7 +11,7 @@ export default class TeamsService {
    */
   static async getAllTeams(): Promise<Team[]> {
     const teams = await prisma.team.findMany(teamQueryArgs);
-    return teams.map(teamsTransformer);
+    return teams.map(teamTransformer);
   }
 
   /**
@@ -30,6 +30,6 @@ export default class TeamsService {
       throw new NotFoundException('Team', teamId);
     }
 
-    return teamsTransformer(team);
+    return teamTransformer(team);
   }
 }
